Merge branch 'development' of https://github.com/lumenlearning/IMathAS into development

+++ models/Exceptions.php
@@ -47,7 +47,9 @@ class Exceptions extends BaseImasExceptions
     public static function deleteExceptionById($id)
     {
         $exception = Exceptions::getById($id);
+        if($exception){
         $exception->delete();
+        }
     }
 
 }
